+++ app/controllers/index_controller.rb
@@ -1,12 +1,6 @@
 require 'csv'
 
 class IndexController < ApplicationController
-  MAX_LENGTH = 1_000
-
-  def limit_length
-    params[:length] = [params[:length].to_i, MAX_LENGTH].min
-  end
-
   def index
     datatable = create_datatable
     respond_to do |format|
@@ -17,13 +11,11 @@ class IndexController < ApplicationController
         }
       }
       format.json {
-        limit_length
-        render json: datatable.serverside.as_json
+        render json: datatable.serverside.limit(params[:length], params[:offset]).as_json
       }
       format.csv {
-        limit_length
         csv = CSV.generate(headers: datatable.column_names, write_headers: true) do |c|
-          datatable.serverside.limit(MAX_LENGTH).as_json.each { |row|  c << row }
+          datatable.serverside.limit.as_json.each { |row|  c << row }  ## TODO
         end
         send_data csv, filename: "#{controller_name}.csv"
       }
@@ -32,7 +24,7 @@ class IndexController < ApplicationController
 
   ## json index access
   def list
-    datatable = create_datatable.serverside.decorate
+    datatable = create_datatable.serverside.paging.decorate
 
     render json:  {
       iTotalRecords:        datatable.records.count,