+++ app/datatables/element_judge_details_datatable.rb
@@ -2,14 +2,7 @@ class ElementJudgeDetailsDatatable < IndexDatatable
   def initialize(*)
     super
     columns([:score_name, :skater_name, :element_number, :element_name, :value, :average, :number, :panel_name])
-
-    columns.sources = {
-      score_name:     'scores.name',
-      skater_name:    'skaters.name',
-      element_number: 'elements.number',
-      element_name:   'elements.name',
-      panel_name:     'panels.name',
-    }
+    columns.sources = source_mappings
   end
 
   def fetch_records