+++ app/parsers/competition_parser/official_parser.rb
@@ -2,7 +2,7 @@ class CompetitionParser
   class OfficialParser < Parser
     def parse(url, category, segment)
       page = get_url(url, read_option: 'r:iso-8859-1').presence || (return [])
-      debug("-- parse panel: #{url}", indent: 3)
+      debug("-- parsing officials: #{url}", indent: 3)
       func = "contains(text(), 'Function')"
       elem = page.xpath("//th[#{func}] | //td[#{func}]") || raise('no Function cell')
       rows = elem.xpath('ancestor::table[1]//tr')