+++ lib/ajax_datatables/datatable/serverside.rb
@@ -10,7 +10,7 @@ module AjaxDatatables::Datatable::Serversidable
   ################
   ## for server-side ajax
   def manipulate(records)
-    super(records).where(build_conditions(columns_searching_nodes)).order(sorting_sql).page(page).per(per)
+    super(records).where(build_conditions(columns_searching_nodes)).order(serverside_ordering_sql)
   end
 
   ################
@@ -26,22 +26,12 @@ module AjaxDatatables::Datatable::Serversidable
 
   ################
   ## sorting
-  def sorting_sql
-    return '' if params[:order].blank?
+  def serverside_ordering_sql
+    return nil if params[:order].blank?
 
     params.require(:order).values.reject { |d| d[:orderable] == 'false' }.map do |item|
       [columns[item[:column].to_i].source, item[:dir]].join(' ')
     end
-  end
-
-  ################
-  ## paging
-  def page
-    params[:start].to_i / per + 1
-  end
-
-  def per
-    (params[:length].to_i.positive?) ? params[:length].to_i : 10
   end
 
 ################