+++ lib/http_get.rb
@@ -14,7 +14,7 @@ module HttpGet
   rescue OpenURI::HTTPError, Errno::ETIMEDOUT, SocketError, Timeout::Error => err
     # #http://www.kraso.sk/wp-content/uploads/sutaze/2014_2015/20141001_ont/html/CAT003RS.HTM returns 404 somehow
     Rails.logger.warn(err.message)
-    debug(err.message)
+    debug("#{err.message}: #{url}")
     nil
   else
     Nokogiri::HTML(html)